Feat: Add math logic, make it functional.
Added logic to the calculator, now it can properly
add, subtract , multiply and divide easily.

Achieved this by first, changing the text field
into an input field then getting an expression
and storing in in a variable. The function
clickHandler check the value if it is a number or
an operator. If the value is operator the previous
expression is stored in a variable called
firstNumber and the operator in operator likewise.
if the operator is an = then call operation
function with the secondNumber as expression.